完成实体类创建，消息发送请求传输对象定义与创建。完成消息处理器的抽象与具体实现

+++ peanut-app-im-core-server/src/main/java/com/sici/chat/im/core/server/common/ChannelHandlerContextCache.java
@@ -30,7 +30,7 @@ public class ChannelHandlerContextCache {
 
     private static Map<Long, ChannelHandlerContext> channelHandlerContextCacheMap = new HashMap<>();
 
-    public static ChannelHandlerContext get(Long userId) {
+    public static ChannelHandlerContext get(Integer userId) {
         return channelHandlerContextCacheMap.get(userId);
     }
